@@ -1,12 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 
-import { Observable } from 'rxjs/Observable';
-import { catchError } from 'rxjs/operators';
-import 'rxjs/add/observable/throw';
-import { of } from 'rxjs/observable/of';
-
-import { environment } from '../../environments/environment';
+import { Observable, of } from 'rxjs';
 
 @Injectable()
 export class ToppingsService {
